create reusable function for piechart
